requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.216.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
sqlalchemy==2.0.23
sqlalchemy[asyncio]==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
requests==2.31.0